move variation to a new component
